# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el panel del bot de citas de una estética
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Service {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  durationMinutes: number;
  category: string;
  active: boolean;
}

export interface ServiceInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 1 */
  durationMinutes: number;
  category: string;
  active?: boolean;
}

export interface ServiceUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 1 */
  durationMinutes?: number;
  category?: string;
  active?: boolean;
}

export interface Faq {
  id: number;
  serviceId: number | null;
  serviceName: string | null;
  question: string;
  answer: string;
  priority: number;
  active: boolean;
}

export interface FaqInput {
  serviceId?: number | null;
  /** @minLength 1 */
  question: string;
  /** @minLength 1 */
  answer: string;
  /** @minimum 0 */
  priority?: number;
  active?: boolean;
}

export interface FaqUpdate {
  serviceId?: number | null;
  /** @minLength 1 */
  question?: string;
  /** @minLength 1 */
  answer?: string;
  /** @minimum 0 */
  priority?: number;
  active?: boolean;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface Appointment {
  id: number;
  clientName: string;
  phone: string;
  serviceId: number;
  serviceName: string;
  area: string;
  scheduledDate: string;
  scheduledTime: string;
  status: AppointmentStatus;
  notes: string;
  createdAt: string;
}

export interface AppointmentInput {
  /** @minLength 1 */
  clientName: string;
  /** @minLength 7 */
  phone: string;
  serviceId: number;
  area: string;
  scheduledDate: string;
  scheduledTime: string;
  status?: AppointmentStatus;
  notes?: string;
}

export interface AppointmentUpdate {
  /** @minLength 1 */
  clientName?: string;
  /** @minLength 7 */
  phone?: string;
  serviceId?: number;
  area?: string;
  scheduledDate?: string;
  scheduledTime?: string;
  status?: AppointmentStatus;
  notes?: string;
}

export interface AvailabilitySlot {
  time: string;
  available: boolean;
}

export interface Dashboard {
  businessName: string;
  todayAppointments: number;
  upcomingAppointments: number;
  activeServices: number;
  faqCount: number;
  botConnected: boolean;
  nextAppointment: Appointment | null;
}

export interface BotStatus {
  connected: boolean;
  provider: string;
  label: string;
  webhookReady: boolean;
}

/**
 * Solicitud inválida
 */
export type BadRequestResponse = Error;

/**
 * Recurso no encontrado
 */
export type NotFoundResponse = Error;

export type ListFaqsParams = {
serviceId?: number;
};

export type ListAppointmentsParams = {
from?: string;
to?: string;
status?: AppointmentStatus;
};

export type GetAvailabilityParams = {
date: string;
serviceId: number;
};