# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el panel del bot de citas de una estética
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from './appointmentStatus';

export interface AppointmentUpdate {
  /** @minLength 1 */
  clientName?: string;
  /** @minLength 7 */
  phone?: string;
  serviceId?: number;
  area?: string;
  scheduledDate?: Date;
  scheduledTime?: string;
  status?: AppointmentStatus;
  notes?: string;
}