# lib/api-zod/src/generated/types/faqInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el panel del bot de citas de una estética
 * OpenAPI spec version: 0.1.0
 */

export interface FaqInput {
  serviceId?: number | null;
  /** @minLength 1 */
  question: string;
  /** @minLength 1 */
  answer: string;
  /** @minimum 0 */
  priority?: number;
  active?: boolean;
}