# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el panel del bot de citas de una estética
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  durationMinutes: number;
  category: string;
  active: boolean;
}