# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para el panel del bot de citas de una estética
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 1 */
  durationMinutes?: number;
  category?: string;
  active?: boolean;
}